refactor: extract logging test support
